Break cli → dataentry import dependency
Extract config types, validation, and ConfigFile constant into
internal/dataentryconfig so cli/validate.go and cli/migrate.go
no longer import the heavy internal/dataentry package.
